scripts/taskfile-validator/pyproject.toml
[project]
name = "taskfile-validator"
version = "0.1.0"
description = "Validate Taskfile.yml against Rule 090 best practices"
requires-python = ">=3.12"
dependencies = ["structlog>=24.0.0", "ruamel.yaml>=0.18"]

[project.optional-dependencies]
test = ["pytest>=8.0.0", "pytest-cov>=4.1.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
